requirements.txt
Django==4.2.0
sentence-transformers>=2.7.0
faiss-cpu==1.13.2
python-dotenv==1.0.0
requests==2.31.0
PyPDF2==3.0.1
